zum Commit vorgemerkte Änderungen: 	geändert:       VUMPS/VumpsPivotStuff.h 	geändert:       VUMPS/VumpsTransferMatrix.h 	geändert:       bases/FermionBaseSU2xSU2.h 	geändert:       models/HubbardSU2xSU2.h 	geändert:       models/HubbardU1xU1.h 	geändert:       programs/test_Heisenberg/test_Heisenberg.cc 	geändert:       programs/test_Hubbard/test_Hubbard.cc 	geändert:       symmetry/S1xS2.h
- Bugfixes bzgl. der neuen Symmetrie-Kombinationen

+++ programs/test_Hubbard/test_Hubbard.cc
@@ -99,7 +99,7 @@ int main (int argc, char* argv[])
 	VERB = static_cast<DMRG::VERBOSITY::OPTION>(args.get<int>("VERB",2));
 	i0 = args.get<int>("i0",L/2);
 	dt = 0.2;
-	size_t V = L*Ly; size_t Vsq = V*V;
+	int V = L*Ly; int Vsq = V*V;
 	
 	Dinit  = args.get<int>("Dmin",2);
 	Dlimit = args.get<int>("Dmax",100);
